+++ android/keyple-plugin/android-nfc/src/main/kotlin/org/eclipse/keyple/plugin/android/nfc/AndroidNfcReaderPreNImpl.kt
@@ -18,4 +18,4 @@ import org.eclipse.keyple.core.plugin.reader.WaitForCardRemovalNonBlocking
  *
  * It uses a Ping monitoring job to detect card removal
  */
-internal object AndroidNfcReaderPreNImpl : AbstractAndroidNfcReader(), WaitForCardRemovalNonBlocking
+internal object AndroidNfcReaderPreNImpl : AndroidNfcReader(), WaitForCardRemovalNonBlocking
